Runtime/RTAcpi*: Updates to the ASL -> AML compiler, it is now possible to produce a valid AML table from the pre processed vbox-cpuhotplug.dsl SSDT, bugref:10733 [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@108091 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/acpi/acpi.cpp
@@ -504,7 +504,7 @@ static void rtAcpiTblAppendNameString(PRTACPITBLINT pThis, const char *pszName)
             cSegments++;
     }
 
-    size_t cbReq = cSegments * 4 * sizeof(uint8_t);
+    uint32_t cbReq = cSegments * 4 * sizeof(uint8_t);
     if (cSegments == 2)
         cbReq++; /* DualName Prefix */
     else if (cSegments != 1)
